fix: Move getNextNonce into ValtBase
