+++ packages/engine/src/vaults/impl/evm/Vault.ts
@@ -336,13 +336,13 @@ export default class Vault extends VaultBase {
         transferInfo.token ?? '',
       );
       if (!token) {
-        throw new Error(`Token not found: ${transferInfo.token as string}`);
+        throw new Error(`Token not found: ${ transferInfo.token as string }`);
       }
       const amountHex = toBigIntHex(amountBN.shiftedBy(token.decimals));
 
-      const data = `${Erc20MethodSelectors.tokenTransfer}${defaultAbiCoder
+      const data = `${ Erc20MethodSelectors.tokenTransfer }${ defaultAbiCoder
         .encode(['address', 'uint256'], [transferInfo.to, amountHex])
-        .slice(2)}`; // method_selector(transfer) + byte32_pad(address) + byte32_pad(value)
+        .slice(2) }`; // method_selector(transfer) + byte32_pad(address) + byte32_pad(value)
       // erc20 token transfer
       return {
         from: transferInfo.from,
@@ -371,7 +371,7 @@ export default class Vault extends VaultBase {
       this.validateAddress(approveInfo.spender),
     ]);
     if (typeof token === 'undefined') {
-      throw new Error(`Token not found: ${approveInfo.token}`);
+      throw new Error(`Token not found: ${ approveInfo.token }`);
     }
 
     const amountBN = new BigNumber(approveInfo.amount);
@@ -382,9 +382,9 @@ export default class Vault extends VaultBase {
     );
     // keccak256(Buffer.from('approve(address,uint256)') => '0x095ea7b3...'
     const methodID = Erc20MethodSelectors.tokenApprove;
-    const data = `${methodID}${defaultAbiCoder
+    const data = `${ methodID }${ defaultAbiCoder
       .encode(['address', 'uint256'], [spender, amountHex])
-      .slice(2)}`;
+      .slice(2) }`;
     return {
       from: approveInfo.from,
       to: approveInfo.token,
@@ -422,9 +422,9 @@ export default class Vault extends VaultBase {
     if (decodedTx.txType === EVMDecodedTxType.TOKEN_TRANSFER) {
       const info = decodedTx.info as EVMDecodedItemERC20Transfer;
       const amountHex = toBigIntHex(amountBN.shiftedBy(info.token.decimals));
-      const data = `${Erc20MethodSelectors.tokenTransfer}${defaultAbiCoder
+      const data = `${ Erc20MethodSelectors.tokenTransfer }${ defaultAbiCoder
         .encode(['address', 'uint256'], [info.recipient, amountHex])
-        .slice(2)}`;
+        .slice(2) }`;
       encodedTx.data = data;
     }
     return encodedTx;
@@ -449,14 +449,14 @@ export default class Vault extends VaultBase {
     } else {
       const amountBN = new BigNumber(amount);
       if (amountBN.isNaN()) {
-        throw new Error(`Invalid amount input: ${amount}`);
+        throw new Error(`Invalid amount input: ${ amount }`);
       }
       amountHex = toBigIntHex(amountBN.shiftedBy(token.decimals));
     }
 
-    const data = `${approveMethodID}${defaultAbiCoder
+    const data = `${ approveMethodID }${ defaultAbiCoder
       .encode(['address', 'uint256'], [spender, amountHex])
-      .slice(2)}`;
+      .slice(2) }`;
     return {
       ...encodedTx,
       data, // Override the data
@@ -572,9 +572,9 @@ export default class Vault extends VaultBase {
       const txData = ethers.utils.serializeTransaction({
         value: encodedTx.value,
         data: encodedTx.data,
-        gasLimit: `0x${(unsignedTx.feeLimit ?? new BigNumber('0')).toString(
+        gasLimit: `0x${ (unsignedTx.feeLimit ?? new BigNumber('0')).toString(
           16,
-        )}`,
+        ) }`,
         to: encodedTx.to,
         chainId: 10, // any number other than 0 will lead to fixed length of data
         gasPrice: '0xf4240', // 0.001 Gwei
@@ -582,9 +582,9 @@ export default class Vault extends VaultBase {
       });
 
       // keccak256(Buffer.from('getL1Fee(bytes)')) => '0x49948e0e...'
-      const data = `0x49948e0e${defaultAbiCoder
+      const data = `0x49948e0e${ defaultAbiCoder
         .encode(['bytes'], [txData])
-        .slice(2)}`;
+        .slice(2) }`;
       const client = await this.getJsonRPCClient();
       const l1FeeHex = await client.rpc.call('eth_call', [
         { to: '0x420000000000000000000000000000000000000F', data },
@@ -614,9 +614,9 @@ export default class Vault extends VaultBase {
         {
           maxPriorityFeePerGas: encodedTx.maxPriorityFeePerGas
             ? this._toNormalAmount(
-                encodedTx.maxPriorityFeePerGas,
-                network.feeDecimals,
-              )
+              encodedTx.maxPriorityFeePerGas,
+              network.feeDecimals,
+            )
             : undefined,
           maxFeePerGas: encodedTx.maxFeePerGas
             ? this._toNormalAmount(encodedTx.maxFeePerGas, network.feeDecimals)
@@ -690,56 +690,6 @@ export default class Vault extends VaultBase {
     return Promise.resolve(encodedTxWithFee);
   }
 
-  private async getNextNonce(
-    networkId: string,
-    dbAccount: DBAccount,
-  ): Promise<number> {
-    const onChainNonce =
-      (
-        await this.engine.providerManager.getAddresses(networkId, [
-          dbAccount.address,
-        ])
-      )[0]?.nonce ?? 0;
-
-    // TODO: Although 100 history items should be enough to cover all the
-    // pending transactions, we need to find a more reliable way.
-    const historyItems = await this.engine.getHistory(
-      networkId,
-      dbAccount.id,
-      undefined,
-      false,
-    );
-    const maxPendingNonce = await simpleDb.history.getMaxPendingNonce({
-      accountId: this.accountId,
-      networkId,
-    });
-    const pendingNonceList = await simpleDb.history.getPendingNonceList({
-      accountId: this.accountId,
-      networkId,
-    });
-    let nextNonce = Math.max(
-      isNil(maxPendingNonce) ? 0 : maxPendingNonce + 1,
-      onChainNonce,
-    );
-    if (Number.isNaN(nextNonce)) {
-      nextNonce = onChainNonce;
-    }
-    if (nextNonce > onChainNonce) {
-      for (let i = onChainNonce; i < nextNonce; i += 1) {
-        if (!pendingNonceList.includes(i)) {
-          nextNonce = i;
-          break;
-        }
-      }
-    }
-
-    if (nextNonce - onChainNonce >= HISTORY_CONSTS.PENDING_QUEUE_MAX_LENGTH) {
-      throw new PendingQueueTooLong(HISTORY_CONSTS.PENDING_QUEUE_MAX_LENGTH);
-    }
-
-    return nextNonce;
-  }
-
   async mmGetPublicKey(options: ISignCredentialOptions): Promise<string> {
     const dbAccount = await this.getDbAccount();
     if (dbAccount.id.startsWith('hd-') || dbAccount.id.startsWith('imported')) {
@@ -797,15 +747,15 @@ export default class Vault extends VaultBase {
 
     if (typeof token === 'undefined') {
       // This will be catched by engine.
-      console.error(`Token not found: ${tokenAddress}`);
+      console.error(`Token not found: ${ tokenAddress }`);
       throw new Error();
     }
 
     // keccak256(Buffer.from('allowance(address,address)') => '0xdd62ed3e...'
     const allowanceMethodID = '0xdd62ed3e';
-    const data = `${allowanceMethodID}${defaultAbiCoder
+    const data = `${ allowanceMethodID }${ defaultAbiCoder
       .encode(['address', 'address'], [dbAccount.address, spenderAddress])
-      .slice(2)}`;
+      .slice(2) }`;
     const client = await this.getJsonRPCClient();
     const rawAllowanceHex = await client.rpc.call('eth_call', [
       { to: token.tokenIdOnNetwork, data },
@@ -821,7 +771,7 @@ export default class Vault extends VaultBase {
       const [encryptedPrivateKey] = Object.values(
         await keyring.getPrivateKeys(password),
       );
-      return `0x${decrypt(password, encryptedPrivateKey).toString('hex')}`;
+      return `0x${ decrypt(password, encryptedPrivateKey).toString('hex') }`;
     }
     throw new OneKeyInternalError(
       'Only credential of HD or imported accounts can be exported',
@@ -996,7 +946,7 @@ export default class Vault extends VaultBase {
         if (
           decodedTx.createdAt &&
           Date.now() - decodedTx.createdAt >
-            HISTORY_CONSTS.SET_IS_FINAL_EXPIRED_IN
+          HISTORY_CONSTS.SET_IS_FINAL_EXPIRED_IN
         ) {
           decodedTx.isFinal = true;
         }
